move `GET /filters` endpoint to supervision controller update test cases
